refactor: merge group authorization and role membership handler
Reduces overhead of the migration to not iterate groups twice.
