+++ migration/identity-migration/src/main/java/io/camunda/migration/identity/handler/sm/GroupMigrationHandler.java
@@ -10,6 +10,7 @@ package io.camunda.migration.identity.handler.sm;
 import static io.camunda.migration.identity.MigrationUtil.convertDecisionPermissions;
 import static io.camunda.migration.identity.MigrationUtil.convertProcessPermissions;
 import static io.camunda.migration.identity.MigrationUtil.normalizeGroupID;
+import static io.camunda.migration.identity.MigrationUtil.normalizeID;
 import static io.camunda.migration.identity.config.saas.StaticEntities.IDENTITY_DECISION_DEFINITION_RESOURCE_TYPE;
 import static io.camunda.migration.identity.config.saas.StaticEntities.IDENTITY_PROCESS_DEFINITION_RESOURCE_TYPE;
 
@@ -22,29 +23,37 @@ import io.camunda.migration.identity.handler.MigrationHandler;
 import io.camunda.security.auth.CamundaAuthentication;
 import io.camunda.service.AuthorizationServices;
 import io.camunda.service.AuthorizationServices.CreateAuthorizationRequest;
+import io.camunda.service.RoleServices;
+import io.camunda.service.RoleServices.RoleMemberRequest;
 import io.camunda.zeebe.protocol.record.value.AuthorizationOwnerType;
 import io.camunda.zeebe.protocol.record.value.AuthorizationResourceType;
+import io.camunda.zeebe.protocol.record.value.EntityType;
 import io.camunda.zeebe.protocol.record.value.PermissionType;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 
-public class GroupAuthorizationMigrationHandler extends MigrationHandler<Group> {
+public class GroupMigrationHandler extends MigrationHandler<Group> {
 
   private final ManagementIdentityClient managementIdentityClient;
+  private final RoleServices roleServices;
   private final AuthorizationServices authorizationServices;
 
   private final AtomicInteger createdGroupAuthorizationCount = new AtomicInteger();
   private final AtomicInteger totalGroupAuthorizationCount = new AtomicInteger();
+  private final AtomicInteger createdGroupRoleMembershipCount = new AtomicInteger();
+  private final AtomicInteger totalGroupRoleMembershipCount = new AtomicInteger();
 
-  public GroupAuthorizationMigrationHandler(
+  public GroupMigrationHandler(
       final ManagementIdentityClient managementIdentityClient,
       final AuthorizationServices authorizationServices,
+      final RoleServices roleServices,
       final CamundaAuthentication authentication,
       final IdentityMigrationProperties migrationProperties) {
     super(migrationProperties.getBackpressureDelay());
     this.managementIdentityClient = managementIdentityClient;
+    this.roleServices = roleServices.withAuthentication(authentication);
     this.authorizationServices = authorizationServices.withAuthentication(authentication);
   }
 
@@ -59,15 +68,47 @@ public class GroupAuthorizationMigrationHandler extends MigrationHandler<Group> 
         group -> {
           final var normalizedGroupId = normalizeGroupID(group);
           createAuthorizationsForGroup(group.id(), normalizedGroupId);
+          createGroupRoleMembership(group);
         });
   }
 
   @Override
   protected void logSummary() {
     logger.info(
-        "Group Authorization migration completed: {} authorizations created out of {} total.",
+        "Group migration completed: {} authorizations created out of {} total, {} role memberships created out of {} total.",
         createdGroupAuthorizationCount.get(),
-        totalGroupAuthorizationCount.get());
+        totalGroupAuthorizationCount.get(),
+        createdGroupRoleMembershipCount,
+        totalGroupRoleMembershipCount);
+  }
+
+  private void createGroupRoleMembership(final Group group) {
+    final var groupRoles = managementIdentityClient.fetchGroupRoles(group.id());
+    totalGroupRoleMembershipCount.addAndGet(groupRoles.size());
+
+    final var normalizedGroupId = normalizeGroupID(group);
+    groupRoles.forEach(
+        role -> {
+          try {
+            final var roleId = normalizeID(role.name());
+            logger.debug("Assigning role '{}' to group '{}'", roleId, normalizedGroupId);
+            final var roleMember =
+                new RoleMemberRequest(roleId, normalizedGroupId, EntityType.GROUP);
+            retryOnBackpressure(
+                () -> roleServices.addMember(roleMember).join(),
+                String.format(
+                    "Failed to assign group '%s' to role '%s'", normalizedGroupId, roleId));
+            createdGroupRoleMembershipCount.incrementAndGet();
+          } catch (final Exception e) {
+            if (!isConflictError(e)) {
+              throw new MigrationException("Failed to assign group: " + normalizedGroupId, e);
+            }
+            logger.debug(
+                "Group with ID '{}' is already assigned to role with ID '{}', skipping assignation.",
+                normalizedGroupId,
+                role.name());
+          }
+        });
   }
 
   private void createAuthorizationsForGroup(final String groupId, final String targetGroupId) {
