+++ migration/identity-migration/src/test/java/io/camunda/migration/identity/handler/sm/SMKeycloakMigrationHandlerConfigTest.java
@@ -32,8 +32,7 @@ public class SMKeycloakMigrationHandlerConfigTest {
     final List<String> expectedOrder =
         List.of(
             "RoleMigrationHandler",
-            "GroupRoleMigrationHandler",
-            "GroupAuthorizationMigrationHandler",
+            "GroupMigrationHandler",
             "UserRoleMigrationHandler",
             "ClientMigrationHandler",
             "AuthorizationMigrationHandler",
